Rename file facade.ts to primitive-valuetypes.ts
