+++ libs/language-server/src/lib/ast/expressions/type-computers/integer-conversion-operator-type-computer.ts
@@ -4,7 +4,7 @@
 
 import { type Valuetype } from '../../wrappers/value-type';
 // eslint-disable-next-line import/no-cycle
-import { PrimitiveValuetypes } from '../../wrappers/value-type/primitive/facade';
+import { PrimitiveValuetypes } from '../../wrappers/value-type/primitive/primitive-valuetypes';
 import { DefaultUnaryOperatorTypeComputer } from '../operator-type-computer';
 
 export class IntegerConversionOperatorTypeComputer extends DefaultUnaryOperatorTypeComputer {
