+++ libs/language-server/src/lib/ast/expressions/type-computers/logical-operator-type-computer.ts
@@ -4,7 +4,7 @@
 
 import { type Valuetype } from '../../wrappers/value-type';
 // eslint-disable-next-line import/no-cycle
-import { PrimitiveValuetypes } from '../../wrappers/value-type/primitive/facade';
+import { PrimitiveValuetypes } from '../../wrappers/value-type/primitive/primitive-valuetypes';
 import { DefaultBinaryOperatorTypeComputer } from '../operator-type-computer';
 
 export class LogicalOperatorTypeComputer extends DefaultBinaryOperatorTypeComputer {
